# LegalConnect.API/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "LegalConnect API",
    "description": "Lawyer Marketplace Platform API",
    "version": "v1"
  },
  "paths": {
...
    "/api/hire-requests": {
      "post": {
        "tags": [
          "HireRequests"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateHireRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/hire-requests/{id}": {
      "get": {
        "tags": [
          "HireRequests"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "HireRequests"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/hire-requests/{id}/accept": {
      "put": {
        "tags": [
          "HireRequests"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/hire-requests/{id}/reject": {
      "put": {
        "tags": [
          "HireRequests"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/hire-requests/{id}/messages": {
      "get": {
        "tags": [
          "HireRequests"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "HireRequests"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SendMessageDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/hire-requests/{id}/messages/read": {
      "put": {
        "tags": [
          "HireRequests"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/hire-requests/deals/{dealId}/proposals": {
      "post": {
        "tags": [
          "HireRequests"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "dealId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateProposalDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/hire-requests/proposals/{proposalId}/accept": {
      "put": {
        "tags": [
          "HireRequests"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "proposalId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RespondToProposalDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/hire-requests/proposals/{proposalId}/reject": {
      "put": {
        "tags": [
          "HireRequests"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "proposalId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RespondToProposalDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/hire-requests/invoices/{invoiceId}/accept": {
      "put": {
        "tags": [
          "HireRequests"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "invoiceId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/hire-requests/invoices/{invoiceId}/paid": {
      "put": {
        "tags": [
          "HireRequests"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "invoiceId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/hire-requests/invoices/{invoiceId}/reject": {
      "put": {
        "tags": [
          "HireRequests"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "invoiceId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateHireRequestDto": {
      "required": [
        "caseType",
        "court",
        "description",
        "lawyerProfileId"
      ],
      "type": "object",
      "properties": {
        "lawyerProfileId": {
          "format": "int32",
          "type": "integer"
        },
        "description": {
          "maxLength": 3000,
          "minLength": 1,
          "type": "string"
        },
        "caseType": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "court": {
          "maxLength": 200,
          "minLength": 1,
          "type": "string"
        },
        "message": {
          "maxLength": 1000,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "CreateProposalDto": {
      "required": [
        "amount",
        "description",
        "title"
      ],
      "type": "object",
      "properties": {
        "title": {
          "maxLength": 200,
          "minLength": 1,
          "type": "string"
        },
        "description": {
          "maxLength": 3000,
          "minLength": 1,
          "type": "string"
        },
        "amount": {
          "format": "double",
          "minimum": 0.01,
          "type": "number"
        },
        "isNegotiable": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
...
    "RespondToProposalDto": {
      "type": "object",
      "properties": {
        "note": {
          "maxLength": 500,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "SendMessageDto": {
      "required": [
        "content"
      ],
      "type": "object",
      "properties": {
        "content": {
          "maxLength": 2000,
          "minLength": 1,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  },
  "securityDefinitions": {
    "Bearer": { }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}